Fix server selector popup position
* Use new CURLOPT_XFERINFOFUNCTION option for progress callback in network client
* Do not show "+" signs in directupload.eu album list

+++ Source/Core/FileDownloader.cpp
@@ -240,7 +240,7 @@ bool CFileDownloader::waitForFinished()
     return true;
 }
 
-int CFileDownloader::ProgressFunc(INetworkClient* userData, double dltotal, double dlnow, double ultotal, double ulnow) {
+int CFileDownloader::ProgressFunc(INetworkClient* userData, int64_t dltotal, int64_t dlnow, int64_t ultotal, int64_t ulnow) {
     if (stopSignal_) {
         return -1;
     }
